feat: S2I Integration (CM-4495) (#45)
* Added builder and basic test

* fix: moved basic_test to tests folder

* fix: minor changes to run scripts

* fix: Added GPU/CPU modes

* fix: added serve command (still does not fix the issue on Sagemaker)

* fix: linting

* fix: generalized base image

* fix: bugfixes to make s2i run on Sagemaker

* fix: linting

* fix: update assemble script to run setup.sh

* fix: changed ReadMe.md

* doc: fixed build -> builder

* lukas/s2i_multiarch (#66)

* feat: allow using different target plattform in s2i base image

feat: use miniconda instead of anaconda to cut down download/image size.

* Add Readme.

* remove curl

* feat: make anaconda install dependant on base image architecture, allows multiarch images in the future

* fix: changing pypi-superai-internal -> pypi-superai for external developers during s2i build

* feat: AI module changes for incorporating s2i (CM-4495) (#64)

* feat: script changes for s2i

* feat: added methods to build and prepare messages

* feat: adding incremental build steps

* refactor: long lines

* feat: changed how model class loading will be performed with s2i changes

* fix: linting dependent

* fix: dockerfile linting

* fix: removed unnecessary test

* style: whitespace

* fix: tests, moved MyKerasModel to different file

* fix: tests, moved individual models to files.

* fix: env file handling and writing. Model name passing through template to script. refactor and fixed test

* fix: added multi layer image creation using s2i. tested with local tests

* style: whitespace

* refactor: extract out base_image

* feat: pull base image and re-tag if it does not exist.

* doc: fixed README.md

* fix: tests. Skipping tests which have ID patching dependency

* fix: reverting docker build_image and _create_dockerfile

* refactor: prepare dependencies

* refactor: moved dockerfile creation

* refactor: optimized imports

* doc: added docstring

* refactor: moved methods to helper

* style: whitespace removed

* fix: passing the correct conda env name as an environment

* style: black

* refactor: renamed scripts_contents.py -> template_contents.py
