+++ superai/meta_ai/s2i/builder/requirements.txt
@@ -1,0 +1,6 @@
+multi-model-server
+sagemaker-inference
+retrying
+superai>=0.2.0.dev16
+superai_schema
+superai-dataclient
