+++ superai/meta_ai/s2i/builder/s2i/bin/run
@@ -1,0 +1,42 @@
+#!/bin/bash -e
+
+function _get_conda_envs() {
+    # Printed value will be list of Conda envs
+    /opt/conda/bin/conda env list | tail -n +3 | head -n -1 | cut -d' ' -f1
+}
+
+function _is_env_present() {
+    # $1 is the Conda env we are probing
+    # Return value will be 0 if present (TRUE)
+
+    # shellcheck disable=SC2155
+    local _conda_envs=$(_get_conda_envs)
+    echo "$_conda_envs" | grep -qw "$1"
+    return $?
+}
+
+#check environment vars
+if [[ -z "$MODEL_NAME" ]]; then
+
+    echo "Failed to find required env vars MODEL_NAME"
+    exit 1
+
+else
+    cd /home/model-server
+
+    if [[ -x before-run ]]; then
+        echo "Executing before-run script"
+        ./before-run
+    fi
+
+    CONDA_ENV_NAME=${CONDA_ENV_NAME:-env}
+    if _is_env_present "$CONDA_ENV_NAME"; then
+        echo "Activating Conda environment '$CONDA_ENV_NAME'"
+        # We need to initialise Conda so that it can run on /bin/sh
+        source /etc/profile.d/conda.sh
+        /opt/conda/bin/conda activate "$CONDA_ENV_NAME"
+    fi
+
+    echo "starting microservice"
+    exec python /home/model-server/dockerd-entrypoint.py "$MODEL_NAME" serve
+fi
