+++ superai/meta_ai/s2i/builder/s2i/bin/save-artifacts
@@ -1,0 +1,12 @@
+#!/bin/sh -e
+
+# For more information see the documentation:
+#	https://github.com/openshift/source-to-image/blob/master/docs/builder_image.md
+#
+#tar cf - <list of files and folder>
+
+#PIP_CACHE=/home/model-server/.pip-cache
+#
+#if [ -d $PIP_CACHE ]; then
+#  tar cf - $PIP_CACHE
+#fi
