+++ superai/meta_ai/s2i/builder/s2i/bin/usage
@@ -1,0 +1,7 @@
+#!/bin/bash -e
+
+cat <<EOF
+This is the superai-s2i-python S2I image:
+To use it, install S2I: https://github.com/openshift/source-to-image
+
+EOF
