+++ tests/meta_ai/mocked_test_ai.py
@@ -2,22 +2,14 @@
 These tests need tensorflow==2.1.0 and opencv-python-headless to run. That is why they are excluded from normal tests.
 """
 
-import json
 import logging
 import os
 import shutil
 import time
-from urllib.request import urlopen
 
-import cv2
-import numpy as np
 import pytest
-import superai_schema.universal_schema.task_schema_functions as df
-import vcr
-from tensorflow import keras
-from tensorflow.keras import layers
-
 import superai
+import vcr
 from superai import DeploymentApiMixin, ModelApiMixin, ProjectAiApiMixin
 from superai.meta_ai import BaseModel
 from superai.meta_ai.ai import AI, Mode, DeployedPredictor, LocalPredictor, AWSPredictor, AITemplate
@@ -27,101 +19,6 @@ from superai.utils import log
 from tests.apis.test_meta_ai import APP_ID, scrub_string, before_record_cb
 
 weights_path = os.path.join(os.path.dirname(__file__), "../../docs/examples/ai/resources/my_model")
-
-
-class MyKerasModel(BaseModel):
-    model = None
-
-    def __init__(self, *args, **kwargs):
-        super(MyKerasModel, self).__init__(*args, **kwargs)
-
-    def load_weights(self, weights_path):
-        self.model = keras.models.load_model(weights_path)
-
-    def predict(self, input):
-        log.info("Predict Input: ", input)
-        image_url = input["data"]["image_url"]
-        req = urlopen(image_url)
-        arr = np.asarray(bytearray(req.read()), dtype=np.uint8)
-        img = cv2.imdecode(arr, cv2.IMREAD_GRAYSCALE)
-        img = cv2.resize(img, (28, 28), interpolation=cv2.INTER_AREA)
-        input = np.reshape(img, (1, 28 * 28))
-        pred = self.model.predict(input)
-        output = np.argmax(pred[0])
-        return [
-            {
-                "prediction": {
-                    "mnist_class": df.exclusive_choice(
-                        choices=list(map(str, range(10))),
-                        selection=int(output),
-                    )
-                },
-                "score": float(pred[0][int(output)]),
-            }
-        ]
-
-    def train(self, model_save_path, **kwargs):
-        (x_train, y_train), (x_test, y_test) = keras.datasets.mnist.load_data()
-
-        # Preprocess the data (these are NumPy arrays)
-        x_train = x_train.reshape(60000, 784).astype("float32") / 255
-        x_test = x_test.reshape(10000, 784).astype("float32") / 255
-
-        y_train = y_train.astype("float32")
-        y_test = y_test.astype("float32")
-
-        # Reserve 10,000 samples for validation
-        x_val = x_train[-10000:]
-        y_val = y_train[-10000:]
-        x_train = x_train[:-10000]
-        y_train = y_train[:-10000]
-
-        model = self.define_model()
-
-        model.compile(
-            optimizer=keras.optimizers.RMSprop(learning_rate=1e-3),
-            loss=keras.losses.SparseCategoricalCrossentropy(),
-            metrics=[keras.metrics.SparseCategoricalAccuracy()],
-        )
-
-        print("Fit model on training data")
-        history = model.fit(
-            x_train,
-            y_train,
-            batch_size=64,
-            epochs=10,
-            # We pass some validation for
-            # monitoring validation loss and metrics
-            # at the end of each epoch
-            validation_data=(x_val, y_val),
-        )
-
-        # Picked from https://www.tensorflow.org/guide/keras/save_and_serialize
-        model.save(model_save_path)
-
-        # we could also store the model config in a json format in the save path
-        json_config = model.to_json()
-        with open(os.path.join(model_save_path, "model_config.json"), "w") as json_writer:
-            json.dump(json_config, json_writer)
-        with open(os.path.join(model_save_path, "config.json"), "w") as json_config_writer:
-            json.dump(kwargs, json_config_writer)
-
-    @staticmethod
-    def define_model():
-        inputs = keras.Input(shape=(784,), name="digits")
-        x = layers.Dense(64, activation="relu", name="dense_1")(inputs)
-        x = layers.Dense(64, activation="relu", name="dense_2")(x)
-        outputs = layers.Dense(10, activation="softmax", name="predictions")(x)
-
-        model = keras.Model(inputs=inputs, outputs=outputs)
-
-        return model
-
-    def to_tf(self):
-        if self.model is not None:
-            return self.model
-        else:
-            return self.define_model()
 
 
 my_vcr = vcr.VCR(
@@ -162,7 +59,7 @@ def ai(cleanup):
         input_schema=Schema(),
         output_schema=Schema(),
         configuration=Config(),
-        model_class=MyKerasModel,
+        model_class="MyKerasModel",
         name="My_template",
         description="Template for my new awesome project",
         requirements=["tensorflow==2.1.0", "opencv-python-headless"],
@@ -174,7 +71,6 @@ def ai(cleanup):
         name="my_mnist_model",
         version=1,
         weights_path=weights_path,
-        model_class=MyKerasModel,
     )
     yield ai
     # delete contents of folder
@@ -190,7 +86,7 @@ def test_create_model(model_api, caplog):
         input_schema=Schema(),
         output_schema=Schema(),
         configuration=Config(),
-        model_class=MyKerasModel,
+        model_class="MyKerasModel",
         name="My_template",
         description="Template for my new awesome project",
         requirements=["tensorflow==2.1.0", "opencv-python-headless"],
@@ -202,7 +98,6 @@ def test_create_model(model_api, caplog):
         name="my_mnist_model",
         version=2,
         weights_path=weights_path,
-        model_class=MyKerasModel,
     )
     assert type(my_ai) == AI
     shutil.rmtree(".AISave")
@@ -236,7 +131,7 @@ def test_mock_load_from_model_hub(ai, monkeypatch):
         ],
     )
     # TODO: replace above patch with a VCS call.
-    monkeypatch.setattr(AI, "load_from_s3", lambda path: AI.load_local(".AISave/" + path.split("s3://")[-1]))
+    monkeypatch.setattr(AI, "load_from_s3", lambda path, other: AI.load_local(".AISave/" + path.split("s3://")[-1]))
 
     ai2 = AI.load("model://my_mnist_model/1")
     assert ai2
@@ -254,8 +149,7 @@ def test_transition_ai_version_stage():
     pass
 
 
-def test_mock_deploy_local(ai, monkeypatch):
-    monkeypatch.setattr(AI, "build_image", lambda *a, **k: None)
+def test_mock_deploy_local(ai):
     predictor = ai.deploy(mode=Mode.LOCAL)
     # standard type checks
     assert type(predictor) == LocalPredictor
@@ -264,7 +158,6 @@ def test_mock_deploy_local(ai, monkeypatch):
 
 @pytest.mark.skip("TODO")
 def test_mock_deploy_sagemaker(ai, monkeypatch):
-    monkeypatch.setattr(AI, "build_image", lambda *a, **k: None)
     monkeypatch.setattr(AI, "push_model", lambda *a, **k: None)
 
     predictor = ai.deploy(mode=Mode.AWS)
@@ -285,7 +178,6 @@ def test_predict(ai):
 
 def test_mock_predict_from_local_deployment(ai, monkeypatch):
     monkeypatch.setattr(LocalPredictor, "predict", ai.predict)
-    monkeypatch.setattr(AI, "build_image", lambda *a, **k: None)
     predictor = ai.deploy(mode=Mode.LOCAL)
 
     assert predictor
@@ -296,9 +188,8 @@ def test_mock_predict_from_local_deployment(ai, monkeypatch):
     assert prediction
 
 
-@pytest.mark.skip("TODO")
+@pytest.mark.skip("TODO: Patching for ID is missing")
 def test_mock_predict_from_sagemaker(ai, monkeypatch):
-    monkeypatch.setattr(AI, "build_model", lambda *a, **k: None)
     monkeypatch.setattr(AI, "push_model", lambda *a, **k: None)
     monkeypatch.setattr(AWSPredictor, "predict", ai.predict)
     predictor = ai.deploy(mode=Mode.AWS)
@@ -311,13 +202,14 @@ def test_mock_predict_from_sagemaker(ai, monkeypatch):
     assert prediction
 
 
-def test_predict_from_sagemaker(cleanup, caplog):
+@pytest.mark.skip("TODO: Patching for ID is missing")
+def test_predict_from_sagemaker(cleanup, caplog, monkeypatch):
     caplog.set_level(logging.INFO)
     template = AITemplate(
         input_schema=Schema(),
         output_schema=Schema(),
         configuration=Config(),
-        model_class=MyKerasModel,
+        model_class="MyKerasModel",
         name="Genre_Template",
         description="Template for genre models",
         requirements=["torch>=1.6"],
@@ -349,7 +241,7 @@ def test_train_and_predict(cleanup):
         input_schema=Schema(),
         output_schema=Schema(),
         configuration=Config(),
-        model_class=MyKerasModel,
+        model_class="MyKerasModel",
         name="My_template",
         description="Template for my new awesome project",
         requirements=["tensorflow", "opencv-python-headless"],
@@ -360,7 +252,6 @@ def test_train_and_predict(cleanup):
         output_params=template.output_schema.parameters(choices=map(str, range(0, 10))),
         name="my_mnist_model",
         version=1,
-        model_class=MyKerasModel,
     )
     model_weights_path = ".AISave/my_model"
     ai.train(model_weights_path, None)
@@ -373,7 +264,6 @@ def test_train_and_predict(cleanup):
         output_params=template.output_schema.parameters(choices=map(str, range(0, 10))),
         name=model_name,
         version=2,
-        model_class=MyKerasModel,
         weights_path=model_weights_path,
     )
     result = my_ai.predict(
@@ -384,17 +274,16 @@ def test_train_and_predict(cleanup):
 
 
 def test_write_dockerfile(ai):
-    ai._create_dockerfile()
+    ai._prepare_dependencies()
     assert os.path.exists(os.path.join(ai._location, "handler.py"))
 
 
-@pytest.mark.skip("Prevent docker in docker")
 def test_build_image():
     template = AITemplate(
         input_schema=Schema(),
         output_schema=Schema(),
         configuration=Config(),
-        model_class=MyKerasModel,
+        model_class="MyKerasModel",
         name="My_template",
         description="Template for my new awesome project",
         requirements=["tensorflow", "opencv-python-headless"],
@@ -408,16 +297,15 @@ def test_build_image():
         name="my_mnist_model",
         version=1,
         weights_path=os.path.join(os.path.dirname(__file__), "../experiments/my_model"),
-        model_class=MyKerasModel,
     )
-    ai._create_dockerfile()
+    ai._prepare_dependencies()
     image_name = "test_build_image"
-    ai.build_image(image_name)
+    ai.build_image_s2i(image_name)
     os.system(f"docker inspect --type=image {image_name}")
 
 
+@pytest.mark.skip("TODO: Patching for ID is missing")
 def test_create_endpoint(ai, monkeypatch):
-    monkeypatch.setattr(AI, "_create_dockerfile", lambda *a, **k: None)
     monkeypatch.setattr(AI, "push_model", lambda *a, **k: None)
     monkeypatch.setattr(DeploymentApiMixin, "check_endpoint_is_available", lambda *a, **k: True)
 
@@ -429,7 +317,8 @@ def test_create_endpoint(ai, monkeypatch):
     assert predictor.predict({"some": "data"}) == 20
 
 
+@pytest.mark.skip("TODO: Patching for ID is missing.")
 def test_delete_endpoint(ai, monkeypatch):
-    monkeypatch.setattr(DeploymentApiMixin, "undeploy", lambda *a, **k: (True, ""))
+    monkeypatch.setattr(DeploymentApiMixin, "undeploy", lambda *a, **k: True)
     res = ai.undeploy()
     assert res
