+++ tests/meta_ai/s2i_test/environment
@@ -1,0 +1,1 @@
+MODEL_NAME=Model
