+++ tests/meta_ai/s2i_test/requirements.txt
@@ -1,0 +1,1 @@
+scikit-learn
