+++ tests/meta_ai/test_unit_ai.py
@@ -3,11 +3,7 @@ import shutil
 import tarfile
 
 import pytest
-
-from superai.meta_ai import AI, BaseModel
-from superai.meta_ai.ai import AITemplate
-from superai.meta_ai.parameters import Config
-from superai.meta_ai.schema import Schema, SchemaParameters
+from superai.meta_ai import AI
 
 
 @pytest.fixture
@@ -41,26 +37,3 @@ def test_compression():
     shutil.rmtree(folder_path)
     shutil.rmtree(another_folder_path)
     os.remove(path_to_tarfile)
-
-
-def test_create_dockerfile_cuda_check(clean):
-    class SomeClass(BaseModel):
-        pass
-
-    mock_template = AITemplate(
-        input_schema=Schema(),
-        output_schema=Schema(),
-        configuration=Config(),
-        model_class=SomeClass,
-        name="MockTemplate",
-        description="Some mocked template",
-    )
-    mock_ai = AI(
-        mock_template, input_params=SchemaParameters(), output_params=SchemaParameters(), name="Mock_AI_instance"
-    )
-
-    content_cpu = mock_ai._create_dockerfile(lambda_mode=False, enable_cuda=False)
-    content_gpu = mock_ai._create_dockerfile(lambda_mode=False, enable_cuda=True)
-
-    assert content_cpu[1] != content_gpu[1], "Different Base images should be present"
-    assert content_cpu[2:] == content_gpu[2:], "Rest of the content should be the same"
